Add files via upload
In this project, we are asked to apply and compare 3 different ML classification approaches to a large-scale dataset using the Iris dataset. We then chose KNN, Naive Bayes, and SVM algorithms as the ML approaches. Comparison between these methods is provided through a Computation-time table, KPI parameters, and Confusion Matrices.